Small refactoring.

+++ Problem1/Problem1.cpp
@@ -5,6 +5,7 @@
 // Date: 13-11-2023
 //
 
+
 #include <iostream>
 #include <vector>
 #include <sstream>
@@ -12,10 +13,12 @@
 
 using namespace std;
 
+
 // Function prototypes
 vector<int> bucketSort(vector<int> unsortedVector);
 string vectorToString(vector<int> vector);
 int findMax(vector<int> v);
+
 
 int main()
 {
@@ -43,7 +46,7 @@ vector<int> bucketSort(vector<int> v)
 	vector<vector<int>> w(max + 1, vector<int>());
 	
 	// Add values to buckets
-	for (int i = 0; i < v.size(); i++)
+	for (int i = 0; i < (int) v.size(); i++)
 	{
 		w[v[i]].push_back(v[i]);
 	}
@@ -52,7 +55,7 @@ vector<int> bucketSort(vector<int> v)
 	vector<int> sorted(0);
 
 	// Append from buckets in order to sorted vector
-	for (int i = 0; i < w.size(); i++)
+	for (int i = 0; i < (int) w.size(); i++)
 	{
 		if (!w[i].empty())
 		{
@@ -65,6 +68,7 @@ vector<int> bucketSort(vector<int> v)
 	return sorted;
 }
 
+
 /*
  * Function: findMax
  * Find the maximum value in a vector
@@ -74,7 +78,7 @@ vector<int> bucketSort(vector<int> v)
 int findMax(vector<int> v)
 {
 	int max = v[0];
-	for (int i = 1; i < v.size(); i++)
+	for (int i = 1; i < (int) v.size(); i++)
 	{
 		if (v[i] > max)
 		{
@@ -95,7 +99,7 @@ int findMax(vector<int> v)
 string vectorToString(vector<int> vector)
 {
 	string output = "{";
-	for (int i = 0; i < vector.size(); i++)
+	for (int i = 0; i < (int) vector.size(); i++)
 	{
 		output += to_string(vector[i]);
 		if (i != vector.size() - 1)
